requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.1
dataclasses==0.6
Django==4.1.7
django-rest-knox==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
pycparser==2.21
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.2.4
tzdata==2022.7
